Rename BitwardenPolicyWarning to BitwardenInfoCalloutCard
